Origin File
This Project is the early Dev of this game from the original prototype. This is a work in progress.

The game as movement and a lerp code w it.

Has tilemaps of a dungeon cave, and a open area w town houses green lakes and so on.

Has two types of ads enemies mole and treant, one boss mod flame chick all have animations

the ads have aniamtion for movement and death

interact function is there

inventory has a shell